bench: Attempt to stabilize benchmarking results

+++ crates/testing/test-programs/fibo/workflow/src/lib.rs
@@ -4,7 +4,7 @@ use crate::bindings::exports::testing::fibo_workflow::workflow::Guest;
 struct Component;
 
 impl Guest for Component {
-    fn fibow(n: u8, iterations: u8) -> u64 {
+    fn fibow(n: u8, iterations: u16) -> u64 {
         let mut last = 0;
         for _ in 0..iterations {
             last = fibo(n);
@@ -12,7 +12,7 @@ impl Guest for Component {
         last
     }
 
-    fn fiboa(n: u8, iterations: u8) -> u64 {
+    fn fiboa(n: u8, iterations: u16) -> u64 {
         let mut last = 0;
         for _ in 0..iterations {
             last = crate::bindings::testing::fibo::fibo::fibo(n);
